adamobile/pyproject.toml
[tool.poetry]
name = "adamobile"
version = "0.1.0"
description = ""
authors = ["dessufalte <72430629+dessufalte@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flet = "^0.25.2"
python-dotenv = "^1.0.1"
plotly = "5.24.1"
matplotlib = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "adamobile.main:main"

